Moved tests of merge-folders to it's own spec file, added a few more tests
